debugage Main application

+++ src/main/java/fr/imie/gmm/entities/base/EntityBase.java
@@ -10,27 +10,27 @@ public abstract class EntityBase {
 	
 	@Id
 	@GeneratedValue(strategy= GenerationType.AUTO)
-	private Integer id;
+	private Long id;
 	
 	public EntityBase() {
 		
 	}
 	
-	public EntityBase(Integer id) {
+	public EntityBase(Long id) {
 		this.id = id;
 	}
 
 	/**
 	 * @return the id
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
